+++ worker/src/admin_api/admin_user_api.ts
@@ -4,7 +4,8 @@ import { CONSTANTS } from '../constants';
 import { getJsonSetting, saveSetting, checkUserPassword, getDomains, getUserRoles } from '../utils';
 import { UserSettings, GeoData, UserInfo } from "../models";
 import { handleListQuery } from '../common'
-import { HonoCustomType } from '../types';
+import UserBindAddressModule from '../user_api/bind_address';
+import i18n from '../i18n';
 
 export default {
     getSetting: async (c: Context<HonoCustomType>) => {
@@ -89,7 +90,8 @@ export default {
     },
     deleteUser: async (c: Context<HonoCustomType>) => {
         const { user_id } = c.req.param();
-        if (!user_id) return c.text("Invalid user_id", 400);
+        const msgs = i18n.getMessagesbyContext(c);
+        if (!user_id) return c.text(msgs.UserNotFoundMsg, 400);
         const { success } = await c.env.DB.prepare(
             `DELETE FROM users WHERE id = ?`
         ).bind(user_id).run();
@@ -104,7 +106,8 @@ export default {
     resetPassword: async (c: Context<HonoCustomType>) => {
         const { user_id } = c.req.param();
         const { password } = await c.req.json();
-        if (!user_id) return c.text("Invalid user_id", 400);
+        const msgs = i18n.getMessagesbyContext(c);
+        if (!user_id) return c.text(msgs.UserNotFoundMsg, 400);
         try {
             checkUserPassword(password);
             const { success } = await c.env.DB.prepare(
@@ -143,5 +146,24 @@ export default {
             return c.text("Failed to update user roles", 500)
         }
         return c.json({ success: true })
-    }
+    },
+    bindAddress: async (c: Context<HonoCustomType>) => {
+        const {
+            user_email, address, user_id, address_id
+        } = await c.req.json();
+        const db_user_id = user_id ?? await c.env.DB.prepare(
+            `SELECT id FROM users WHERE user_email = ?`
+        ).bind(user_email).first<number | undefined | null>("id");
+        const db_address_id = address_id ?? await c.env.DB.prepare(
+            `SELECT id FROM address WHERE name = ?`
+        ).bind(address).first<number | undefined | null>("id");
+        return await UserBindAddressModule.bindByID(c, db_user_id, db_address_id);
+    },
+    getBindedAddresses: async (c: Context<HonoCustomType>) => {
+        const { user_id } = c.req.param();
+        const results = await UserBindAddressModule.getBindedAddressesById(c, user_id);
+        return c.json({
+            results: results,
+        });
+    },
 }