+++ worker/src/mails_api/index.ts
@@ -1,8 +1,8 @@
-import { Hono } from 'hono'
+import { Context, Hono } from 'hono'
 
-import { HonoCustomType } from "../types";
-import { getBooleanValue, getJsonSetting, checkCfTurnstile, getStringValue } from '../utils';
-import { newAddress, handleListQuery, deleteAddressWithData, getAddressPrefix, getAllowDomains } from '../common'
+import i18n from '../i18n';
+import { getBooleanValue, getJsonSetting, checkCfTurnstile, getStringValue, getSplitStringListValue } from '../utils';
+import { newAddress, handleListQuery, deleteAddressWithData, getAddressPrefix, getAllowDomains, updateAddressUpdatedAt } from '../common'
 import { CONSTANTS } from '../constants'
 import auto_reply from './auto_reply'
 import webhook_settings from './webhook_settings';
@@ -16,6 +16,7 @@ api.get('/api/webhook/settings', webhook_settings.getWebhookSettings)
 api.post('/api/webhook/settings', webhook_settings.saveWebhookSettings)
 api.post('/api/webhook/test', webhook_settings.testWebhookSettings)
 api.get('/api/attachment/list', s3_attachment.list)
+api.post('/api/attachment/delete', s3_attachment.deleteKey)
 api.post('/api/attachment/put_url', s3_attachment.getSignedPutUrl)
 api.post('/api/attachment/get_url', s3_attachment.getSignedGetUrl)
 
@@ -25,6 +26,7 @@ api.get('/api/mails', async (c) => {
         return c.json({ "error": "No address" }, 400)
     }
     const { limit, offset } = c.req.query();
+    if (Number.parseInt(offset) <= 0) await updateAddressUpdatedAt(c, address);
     return await handleListQuery(c,
         `SELECT * FROM raw_mails where address = ?`,
         `SELECT count(*) as count FROM raw_mails where address = ?`,
@@ -42,8 +44,10 @@ api.get('/api/mail/:mail_id', async (c) => {
 })
 
 api.delete('/api/mails/:id', async (c) => {
+    const lang = c.get("lang") || c.env.DEFAULT_LANG;
+    const msgs = i18n.getMessages(lang);
     if (!getBooleanValue(c.env.ENABLE_USER_DELETE_EMAIL)) {
-        return c.text("User delete email is disabled", 403)
+        return c.text(msgs.UserDeleteEmailDisabledMsg, 403)
     }
     const { address } = c.get("jwtPayload")
     const { id } = c.req.param();
@@ -59,16 +63,18 @@ api.delete('/api/mails/:id', async (c) => {
 api.get('/api/settings', async (c) => {
     const { address, address_id } = c.get("jwtPayload")
     const user_role = c.get("userRolePayload")
+    const lang = c.get("lang") || c.env.DEFAULT_LANG;
+    const msgs = i18n.getMessages(lang);
     if (address_id && address_id > 0) {
         try {
             const db_address_id = await c.env.DB.prepare(
                 `SELECT id FROM address where id = ? `
             ).bind(address_id).first("id");
             if (!db_address_id) {
-                return c.text("Invalid address", 400)
+                return c.text(msgs.InvalidAddressMsg, 400)
             }
         } catch (error) {
-            return c.text("Invalid address", 400)
+            return c.text(msgs.InvalidAddressMsg, 400)
         }
     }
     // check address id
@@ -78,21 +84,17 @@ api.get('/api/settings', async (c) => {
                 `SELECT id FROM address where name = ? `
             ).bind(address).first("id");
             if (!db_address_id) {
-                return c.text("Invalid address", 400)
+                return c.text(msgs.InvalidAddressMsg, 400)
             }
         }
     } catch (error) {
-        return c.text("Invalid address", 400)
+        return c.text(msgs.InvalidAddressMsg, 400)
     }
-    // update address updated_at
-    try {
-        c.env.DB.prepare(
-            `UPDATE address SET updated_at = datetime('now') where name = ?`
-        ).bind(address).run();
-    } catch (e) {
-        console.warn("Failed to update address")
-    }
-    const is_no_limit_send_balance = user_role && user_role === getStringValue(c.env.NO_LIMIT_SEND_ROLE);
+
+    await updateAddressUpdatedAt(c, address);
+
+    const no_limit_roles = getSplitStringListValue(c.env.NO_LIMIT_SEND_ROLE);
+    const is_no_limit_send_balance = user_role && no_limit_roles.includes(user_role);
     const balance = is_no_limit_send_balance ? 99999 : await c.env.DB.prepare(
         `SELECT balance FROM address_sender where address = ? and enabled = 1`
     ).bind(address).first("balance");
@@ -103,13 +105,15 @@ api.get('/api/settings', async (c) => {
 })
 
 api.post('/api/new_address', async (c) => {
+    const lang = c.get("lang") || c.env.DEFAULT_LANG;
+    const msgs = i18n.getMessages(lang);
     if (getBooleanValue(c.env.DISABLE_ANONYMOUS_USER_CREATE_EMAIL)
         && !c.get("userPayload")
     ) {
-        return c.text("New address for anonymous user is disabled", 403)
+        return c.text(msgs.NewAddressAnonymousDisabledMsg, 403)
     }
     if (!getBooleanValue(c.env.ENABLE_USER_CREATE_EMAIL)) {
-        return c.text("New address is disabled", 403)
+        return c.text(msgs.NewAddressDisabledMsg, 403)
     }
     // eslint-disable-next-line prefer-const
     let { name, domain, cf_token } = await c.req.json();
@@ -117,7 +121,7 @@ api.post('/api/new_address', async (c) => {
     try {
         await checkCfTurnstile(c, cf_token);
     } catch (error) {
-        return c.text("Failed to check cf turnstile", 500)
+        return c.text(msgs.TurnstileCheckFailedMsg, 500)
     }
     // if no name, generate random name
     if (!name) {
@@ -143,7 +147,7 @@ api.post('/api/new_address', async (c) => {
         });
         return c.json(res);
     } catch (e) {
-        return c.text(`Failed create address: ${(e as Error).message}`, 400)
+        return c.text(`${msgs.FailedCreateAddressMsg}: ${(e as Error).message}`, 400)
     }
 })