+++ worker/src/telegram_api/telegram.ts
@@ -5,7 +5,6 @@ import { callbackQuery } from "telegraf/filters";
 
 import { CONSTANTS } from "../constants";
 import { getDomains, getJsonObjectValue, getStringValue } from '../utils';
-import { HonoCustomType, ParsedEmailContext } from "../types";
 import { TelegramSettings } from "./settings";
 import { bindTelegramAddress, deleteTelegramAddress, jwtListToAddressData, tgUserNewAddress, unbindTelegramAddress, unbindTelegramByAddress } from "./common";
 import { commonParseMail } from "../common";
@@ -102,7 +101,10 @@ export function newTelegramBot(c: Context<HonoCustomType>, token: string): Teleg
             const res = await tgUserNewAddress(c, userId.toString(), address);
             return await ctx.reply(`创建地址成功:\n`
                 + `地址: ${res.address}\n`
-                + `凭证: ${res.jwt}\n`
+                + `凭证: \`${res.jwt}\`\n`,
+                {
+                    parse_mode: "Markdown"
+                }
             );
         } catch (e) {
             return await ctx.reply(`创建地址失败: ${(e as Error).message}`);